requirements.txt
Django~=4.2
psycopg2-binary~=2.9
python-dotenv~=1.0
django-autocomplete-light